+++ HISTORY.rst
@@ -8,6 +8,8 @@ Unreleased
 
 **Features and Improvements**
 
+* ALCN-1355: Block product quick create
+
 **Bugfixes**
 
 * ALCN-1454: Fix delivery rounds plan generation error